+++ uefi-raw/CHANGELOG.md
@@ -15,6 +15,8 @@
 - Added `UsbIoProtocol`.
 - Added `Usb2HostControllerProtocol`.
 - Added  `DevicePathProtocol::length()` properly constructing the `u16` value
+- Type `IpAddress` is now tightly integrated with `core::net::IpAddr`, e.g.,
+  various `From` implementations are available.
 
 ## Changed
 - **Breaking:** Types `Ipv4Address` and `Ipv6Address` have been removed. They